pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "groq>=0.31.0",
    "gtts>=2.5.4",
    "openai>=1.99.9",
    "requests>=2.32.4",
    "streamlit>=1.48.0",
]
